board generations spawn jsons

+++ src/data_base/domain_board_index.rs
@@ -1,2 +1,2 @@
 #[derive(Debug)]
-pub struct DomainBoardIndex(pub u8);
+pub struct DomainBoardIndex(pub usize);
